# contracts/web-api.openapi.json
{
  "openapi": "3.1.2",
  "info": {
    "title": "Cherry OJ Web API",
    "version": "0.2.0",
    "description": "浏览器与 Gateway 之间的公开 REST 契约。请求使用 endpoint DTO，普通 JSON 成功响应使用 ApiSuccess，失败响应使用 RFC 9457 Problem Details。 管理部署接口的 503 可返回固定节点错误：NO_ONLINE_JUDGE_NODE, JUDGE_NODE_UNREACHABLE, JUDGE_NODE_DATA_REJECTED, JUDGE_NODE_RECEIPT_MISMATCH。详情由固定文案生成，不透传节点正文；readiness checks 的 ONLINE_JUDGE_NODE 表示 ACTIVE 环境存在在线节点，DEPLOYMENT 仅在该在线节点持有匹配摘要回执时通过。"
  },
...
  "paths": {
...
    "/api/problems": {
      "get": {
        "operationId": "listProblems",
        "summary": "浏览公开题库",
        "security": [],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            }
          },
          {
            "name": "difficulty",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ProblemDifficulty"
            }
          },
          {
            "name": "tag",
            "in": "query",
            "schema": {
              "type": "array",
              "maxItems": 10,
              "items": {
                "$ref": "#/components/schemas/ProblemTag"
              }
            },
            "style": "form",
            "explode": true
          },
          {
            "name": "codeMode",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ProblemCodeMode"
            }
          },
          {
            "name": "language",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/LanguageId"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ProblemSort"
            }
          },
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/CursorPageSize"
          }
        ],
        "responses": {
          "200": {
            "description": "公开题目列表；只包含当前公开版本",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/XRequestId"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemListSuccess"
                },
                "example": {
                  "data": {
                    "items": [
                      {
                        "problemId": "019c8e42-7f70-7000-8000-000000000101",
                        "slug": "a-plus-b",
                        "title": "A+B Problem",
                        "difficulty": "EASY",
                        "tags": [
                          "入门"
                        ],
                        "codeMode": "ACM",
                        "allowedLanguages": [
                          {
                            "id": "cpp",
                            "displayName": "C++"
                          }
                        ]
                      }
                    ]
                  },
                  "meta": {
                    "requestId": "req_01K37XZ3MFXBK92WMG67G4XFN0",
                    "pagination": {
                      "kind": "cursor",
                      "nextCursor": null,
                      "hasMore": false
                    }
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ApiProblemResponse"
          }
        }
      }
    },
    "/api/problems/{slug}": {
      "get": {
        "operationId": "getProblem",
        "summary": "读取公开题目当前版本",
        "security": [],
        "parameters": [
          {
            "$ref": "#/components/parameters/ProblemSlug"
          }
        ],
        "responses": {
          "200": {
            "description": "公开题目详情；不包含隐藏数据、判题模板、作者或审计字段",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/XRequestId"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailSuccess"
                },
                "example": {
                  "data": {
                    "problemId": "019c8e42-7f70-7000-8000-000000000101",
                    "slug": "a-plus-b",
                    "title": "A+B Problem",
                    "difficulty": "EASY",
                    "tags": [
                      "入门"
                    ],
                    "codeMode": "ACM",
                    "statementMarkdown": "计算两个整数之和。",
                    "inputDescriptionMarkdown": "输入两个整数。",
                    "outputDescriptionMarkdown": "输出它们的和。",
                    "constraintsMarkdown": "整数在 32 位有符号范围内。",
                    "hintMarkdown": null,
                    "samples": [
                      {
                        "ordinal": 1,
                        "input": "1 2\n",
                        "output": "3\n",
                        "explanationMarkdown": null
                      }
                    ],
                    "allowedLanguages": [
                      {
                        "id": "cpp",
                        "displayName": "C++",
                        "starterCode": "#include <iostream>\n"
                      }
                    ]
                  },
                  "meta": {
                    "requestId": "req_01K37XZ3MFXBK92WMG67G4XFN0"
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ApiProblemResponse"
          }
        }
      }
    },
...
    "/api/admin/problems/{problemId}": {
      "get": {
        "operationId": "getAdminProblem",
        "summary": "读取题目管理详情",
        "description": "题目只有一份内容：题面、样例、语言模板与测试数据摘要都在这里。",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ProblemId"
          }
        ],
        "responses": {
          "200": {
            "description": "题目管理模型",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/XRequestId"
              },
              "Cache-Control": {
                "required": true,
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProblemSuccess"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ApiProblemResponse"
          }
        }
      },
      "patch": {
        "operationId": "updateProblem",
        "summary": "保存题目",
        "description": "改了就是改了，没有草稿或版本。公开题目的修改立即对学生生效，因此不允许把公开题目改成空题面或没有样例；旧 rowVersion 返回 409 ROW_VERSION_CONFLICT。可见性与测试数据由专门的接口改变。",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ProblemId"
          },
          {
            "$ref": "#/components/parameters/CsrfToken"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProblemRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "题目已保存",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/XRequestId"
              },
              "Cache-Control": {
                "required": true,
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProblemSuccess"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ApiProblemResponse"
          }
        }
      },
      "delete": {
        "operationId": "deleteProblem",
        "summary": "删除从未公开过的题目",
        "description": "只允许删除从未公开过的题目（它不可能有提交）；题面、样例、语言、审计事件和测试数据目录一并删除。公开过的题目只能归档，返回 409 RESOURCE_STATE_CONFLICT。",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ProblemId"
          },
          {
            "$ref": "#/components/parameters/RowVersion"
          },
          {
            "$ref": "#/components/parameters/CsrfToken"
          }
        ],
        "responses": {
          "204": {
            "$ref": "#/components/responses/NoContentResponse"
          },
          "default": {
            "$ref": "#/components/responses/ApiProblemResponse"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "CHERRY_SESSION",
        "description": "浏览器随机 Session Cookie；不是内部 JWT。"
      }
    },
    "parameters": {
      "CsrfToken": {
        "name": "X-CSRF-Token",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 16,
          "maxLength": 512
        }
      },
...
      "ProblemId": {
        "name": "problemId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
      "ProblemSlug": {
        "name": "slug",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "#/components/schemas/ProblemSlug"
        }
      },
      "Cursor": {
        "name": "cursor",
        "in": "query",
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 2048
        },
        "description": "服务端签发的不透明游标；非法或失效游标返回 400 INVALID_CURSOR。"
      },
      "CursorPageSize": {
        "name": "size",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20
        }
      },
      "RowVersion": {
        "name": "rowVersion",
        "in": "query",
        "required": true,
        "schema": {
          "type": "integer",
          "minimum": 0
        },
        "description": "乐观锁版本；过期值返回 409 ROW_VERSION_CONFLICT。"
      }
    },
    "headers": {
      "XRequestId": {
        "description": "Gateway 生成的公开请求标识；必须与响应 body 的 meta.requestId 相同。",
        "required": true,
        "schema": {
          "$ref": "#/components/schemas/RequestId"
        }
      }
    },
    "responses": {
      "NoContentResponse": {
        "description": "操作成功且无响应 body。",
        "headers": {
          "X-Request-Id": {
            "$ref": "#/components/headers/XRequestId"
          },
          "Cache-Control": {
            "required": true,
            "schema": {
              "const": "no-store",
              "type": "string"
            }
          }
        }
      },
      "ApiProblemResponse": {
        "description": "符合 RFC 9457 且经过 Gateway 脱敏的错误响应。实际 HTTP status 与 body.status 相同。",
        "headers": {
          "X-Request-Id": {
            "$ref": "#/components/headers/XRequestId"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ApiProblem"
            },
            "example": {
              "type": "urn:cherry-oj:problem:validation-failed",
              "title": "请求参数校验失败",
              "status": 422,
              "detail": "请检查标记字段。",
              "instance": "urn:cherry-oj:request:req_01K37XZ3MFXBK92WMG67G4XFN0",
              "code": "VALIDATION_FAILED",
              "meta": {
                "requestId": "req_01K37XZ3MFXBK92WMG67G4XFN0"
              },
              "violations": [
                {
                  "path": "title",
                  "code": "SIZE",
                  "message": "标题长度必须在 1 到 120 之间。"
                }
              ]
            }
          }
        }
      }
    },
    "schemas": {
...
      "RequestId": {
        "type": "string",
        "pattern": "^req_[A-Za-z0-9_-]{16,64}$",
        "description": "Gateway 生成的 opaque public request ID，不等同于 session、trace 或幂等键。"
      },
      "ApiMeta": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "requestId"
        ],
        "properties": {
          "requestId": {
            "$ref": "#/components/schemas/RequestId"
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        },
        "examples": [
          {
            "requestId": "req_01K37XZ3MFXBK92WMG67G4XFN0"
          }
        ]
      },
      "Pagination": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/CursorPagination"
          },
          {
            "$ref": "#/components/schemas/PagePagination"
          }
        ],
        "discriminator": {
          "propertyName": "kind",
          "mapping": {
            "cursor": "#/components/schemas/CursorPagination",
            "page": "#/components/schemas/PagePagination"
          }
        }
      },
...
      "FieldViolation": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "path",
          "code",
          "message"
        ],
        "properties": {
          "path": {
            "type": "string",
            "minLength": 1,
            "maxLength": 512
          },
          "code": {
            "type": "string",
            "pattern": "^[A-Z][A-Z0-9_]{0,63}$"
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1024
          }
        },
        "examples": [
          {
            "path": "title",
            "code": "SIZE",
            "message": "标题长度必须在 1 到 120 之间。"
          }
        ]
      },
      "ApiProblem": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "type",
          "title",
          "status",
          "code",
          "meta"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri-reference"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 256
          },
          "status": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "detail": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2048
          },
          "instance": {
            "type": "string",
            "format": "uri-reference"
          },
          "code": {
            "type": "string",
            "pattern": "^[A-Z][A-Z0-9_]{0,63}$"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          },
          "violations": {
            "type": "array",
            "maxItems": 100,
            "items": {
              "$ref": "#/components/schemas/FieldViolation"
            }
          }
        },
        "examples": [
          {
            "type": "urn:cherry-oj:problem:validation-failed",
            "title": "请求参数校验失败",
            "status": 422,
            "detail": "请检查标记字段。",
            "instance": "urn:cherry-oj:request:req_01K37XZ3MFXBK92WMG67G4XFN0",
            "code": "VALIDATION_FAILED",
            "meta": {
              "requestId": "req_01K37XZ3MFXBK92WMG67G4XFN0"
            },
            "violations": [
              {
                "path": "title",
                "code": "SIZE",
                "message": "标题长度必须在 1 到 120 之间。"
              }
            ]
          }
        ]
      },
...
      "ProblemSlug": {
        "type": "string",
        "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
        "minLength": 1,
        "maxLength": 128
      },
      "ProblemTag": {
        "type": "string",
        "minLength": 1,
        "maxLength": 32
      },
      "LanguageId": {
        "type": "string",
        "pattern": "^[a-z][a-z0-9-]{0,31}$"
      },
      "ProblemDifficulty": {
        "type": "string",
        "enum": [
          "UNRATED",
          "EASY",
          "MEDIUM",
          "HARD"
        ]
      },
      "ProblemCodeMode": {
        "type": "string",
        "enum": [
          "ACM",
          "CORE"
        ]
      },
      "ProblemSort": {
        "type": "string",
        "enum": [
          "UPDATED_DESC",
          "UPDATED_ASC",
          "TITLE_ASC"
        ],
        "default": "UPDATED_DESC"
      },
      "ProblemVisibility": {
        "type": "string",
        "enum": [
          "PRIVATE",
          "PUBLIC"
        ]
      },
      "ProblemStatus": {
        "type": "string",
        "enum": [
          "ACTIVE",
          "ARCHIVED"
        ]
      },
...
      "PublicLanguageSummary": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "displayName"
        ],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/LanguageId"
          },
          "displayName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          }
        }
      },
      "PublicProblemLanguage": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "displayName",
          "starterCode"
        ],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/LanguageId"
          },
          "displayName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          },
          "starterCode": {
            "type": "string",
            "maxLength": 1048576
          }
        }
      },
      "ProblemSample": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "ordinal",
          "input",
          "output",
          "explanationMarkdown"
        ],
        "properties": {
          "ordinal": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          },
          "input": {
            "type": "string",
            "maxLength": 1048576
          },
          "output": {
            "type": "string",
            "maxLength": 1048576
          },
          "explanationMarkdown": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 1048576
          }
        }
      },
      "ProblemSummary": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "problemId",
          "slug",
          "title",
          "difficulty",
          "tags",
          "codeMode",
          "allowedLanguages"
        ],
        "properties": {
          "problemId": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "$ref": "#/components/schemas/ProblemSlug"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 512
          },
          "difficulty": {
            "$ref": "#/components/schemas/ProblemDifficulty"
          },
          "tags": {
            "type": "array",
            "maxItems": 20,
            "uniqueItems": true,
            "items": {
              "$ref": "#/components/schemas/ProblemTag"
            }
          },
          "codeMode": {
            "$ref": "#/components/schemas/ProblemCodeMode"
          },
          "allowedLanguages": {
            "type": "array",
            "minItems": 1,
            "maxItems": 20,
            "items": {
              "$ref": "#/components/schemas/PublicLanguageSummary"
            }
          }
        }
      },
      "ProblemListData": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "maxItems": 100,
            "items": {
              "$ref": "#/components/schemas/ProblemSummary"
            }
          }
        }
      },
      "ProblemListSuccess": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data",
          "meta"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ProblemListData"
          },
          "meta": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ApiMeta"
              }
            ],
            "required": [
              "requestId",
              "pagination"
            ]
          }
        }
      },
      "ProblemDetail": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "problemId",
          "slug",
          "codeMode",
          "title",
          "difficulty",
          "tags",
          "statementMarkdown",
          "inputDescriptionMarkdown",
          "outputDescriptionMarkdown",
          "constraintsMarkdown",
          "hintMarkdown",
          "samples",
          "allowedLanguages"
        ],
        "properties": {
          "problemId": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "$ref": "#/components/schemas/ProblemSlug"
          },
          "codeMode": {
            "$ref": "#/components/schemas/ProblemCodeMode"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 512
          },
          "difficulty": {
            "$ref": "#/components/schemas/ProblemDifficulty"
          },
          "tags": {
            "type": "array",
            "maxItems": 20,
            "uniqueItems": true,
            "items": {
              "$ref": "#/components/schemas/ProblemTag"
            }
          },
          "statementMarkdown": {
            "type": "string",
            "maxLength": 16777215
          },
          "inputDescriptionMarkdown": {
            "type": "string",
            "maxLength": 16777215
          },
          "outputDescriptionMarkdown": {
            "type": "string",
            "maxLength": 16777215
          },
          "constraintsMarkdown": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 16777215
          },
          "hintMarkdown": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 16777215
          },
          "samples": {
            "type": "array",
            "maxItems": 100,
            "items": {
              "$ref": "#/components/schemas/ProblemSample"
            }
          },
          "allowedLanguages": {
            "type": "array",
            "minItems": 1,
            "maxItems": 20,
            "items": {
              "$ref": "#/components/schemas/PublicProblemLanguage"
            }
          }
        }
      },
      "ProblemDetailSuccess": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data",
          "meta"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ProblemDetail"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        }
      },
      "AdminProblem": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "slug",
          "visibility",
          "status",
          "codeMode",
          "title",
          "statementMarkdown",
          "inputDescriptionMarkdown",
          "outputDescriptionMarkdown",
          "constraintsMarkdown",
          "hintMarkdown",
          "difficulty",
          "tags",
          "samples",
          "allowedLanguages",
          "testData",
          "createdAt",
          "updatedAt",
          "publishedAt",
          "rowVersion"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "$ref": "#/components/schemas/ProblemSlug"
          },
          "visibility": {
            "$ref": "#/components/schemas/ProblemVisibility"
          },
          "status": {
            "$ref": "#/components/schemas/ProblemStatus"
          },
          "codeMode": {
            "const": "ACM",
            "type": "string"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 512
          },
          "statementMarkdown": {
            "type": "string",
            "maxLength": 16777215
          },
          "inputDescriptionMarkdown": {
            "type": "string",
            "maxLength": 16777215
          },
          "outputDescriptionMarkdown": {
            "type": "string",
            "maxLength": 16777215
          },
          "constraintsMarkdown": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 16777215
          },
          "hintMarkdown": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 16777215
          },
          "difficulty": {
            "$ref": "#/components/schemas/ProblemDifficulty"
          },
          "tags": {
            "type": "array",
            "maxItems": 20,
            "uniqueItems": true,
            "items": {
              "$ref": "#/components/schemas/ProblemTag"
            }
          },
          "samples": {
            "type": "array",
            "maxItems": 100,
            "items": {
              "$ref": "#/components/schemas/ProblemSample"
            }
          },
          "allowedLanguages": {
            "type": "array",
            "minItems": 1,
            "maxItems": 1,
            "items": {
              "$ref": "#/components/schemas/AdminProblemLanguage"
            }
          },
          "testData": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TestData"
              },
              {
                "type": "null"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "publishedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "rowVersion": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "AdminProblemSuccess": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data",
          "meta"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AdminProblem"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        }
      },
...
      "UpdateProblemRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "slug",
          "title",
          "statementMarkdown",
          "inputDescriptionMarkdown",
          "outputDescriptionMarkdown",
          "constraintsMarkdown",
          "hintMarkdown",
          "difficulty",
          "tags",
          "samples",
          "starterCode",
          "rowVersion"
        ],
        "properties": {
          "slug": {
            "$ref": "#/components/schemas/ProblemSlug"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 512
          },
          "statementMarkdown": {
            "type": "string",
            "maxLength": 16777215
          },
          "inputDescriptionMarkdown": {
            "type": "string",
            "maxLength": 16777215
          },
          "outputDescriptionMarkdown": {
            "type": "string",
            "maxLength": 16777215
          },
          "constraintsMarkdown": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 16777215
          },
          "hintMarkdown": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 16777215
          },
          "difficulty": {
            "$ref": "#/components/schemas/ProblemDifficulty"
          },
          "tags": {
            "type": "array",
            "maxItems": 20,
            "uniqueItems": true,
            "items": {
              "$ref": "#/components/schemas/ProblemTag"
            }
          },
          "samples": {
            "type": "array",
            "maxItems": 100,
            "items": {
              "$ref": "#/components/schemas/ProblemSample"
            }
          },
          "starterCode": {
            "type": "string",
            "maxLength": 1048576
          },
          "rowVersion": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
      "AdminProblemLanguage": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "displayName",
          "starterCode"
        ],
        "properties": {
          "id": {
            "const": "cpp",
            "type": "string"
          },
          "displayName": {
            "const": "C++",
            "type": "string"
          },
          "starterCode": {
            "type": "string",
            "maxLength": 1048576
          }
        }
      },
...
      "TestData": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "digest",
          "caseCount",
          "totalBytes",
          "updatedAt",
          "manifest"
        ],
        "properties": {
          "digest": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$",
            "description": "testdata.json 的 digest：所有文件的 SHA-256 按固定格式再哈希，内容变则指纹变。"
          },
          "caseCount": {
            "type": "integer",
            "minimum": 1,
            "maximum": 1000
          },
          "totalBytes": {
            "type": "integer",
            "minimum": 0
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "manifest": {
            "$ref": "#/components/schemas/TestDataManifest"
          }
        }
      },
...
    }
  }
}